Also hide balance in wallet
